+++ CHANGELOG.md
@@ -1,4 +1,19 @@
-﻿## Version 4.12 [06-05-2026]
+﻿## Version 4.13 [13-05-2026]
+
+**All WoW Versions:**  
+- Added option to disable and hide the UI Errors Frame to /lucky config  
+- Updated deDE locales  
+- Updated frFR locales (Credits: Oniria)  
+
+**Retail:**  
+- Added option to auto accept role check popups in LFG to /lucky config  
+- Added option to enable double-clicking to signup in LFG to /lucky config  
+- Added skin for the Blizzard Cooldown Manager settings panel  
+- Removed BetterCooldownManager profile  
+- Updated SkironCooldownManager profile for all 40 specs (Requires alpha)  
+- Updated the edit mode specialization strings for all 40 specs (Advanced)  
+
+## Version 4.12 [06-05-2026]
 
 **All WoW Versions:**  
 - 1080p improvements (BigWigs and Details)